ci: add release step

+++ .github/actions/build_deb.sh
@@ -39,15 +39,16 @@ if [ -n "${CI_RUN+1}" ] ; then
     echo "silkit_debian_revision=${SILKIT_DEBIAN_REVISION}" >> "$GITHUB_OUTPUT"
 fi
 
-
 ## Set GIT Options
 if [ -n SILKIT_VENDORED_PACKAGES ] ; then
-   SUBMODULE_CMD="--recurse-submodules --shallow-submodules"
+    echo "Vendoring Packages!"
+    SUBMODULE_CMD="--recurse-submodules --shallow-submodules"
 else
-   SUBMODULE_CMD=""
+    SUBMODULE_CMD=""
 fi
 
-
+# Since we want to allow building SIL KIT from VERSIONS different from the current CHANGELOG Version
+# We need to distinguish how to checkout and reset the local git repo
 if [ -n SILKIT_REVISION ] ; then
     CLONE_VERSION="main"
 else
@@ -66,12 +67,23 @@ if [ ! -d ./libsilkit-${SILKIT_VERSION} ]; then
     exit 64
 fi
 
-echo "SILKIT REVISION: ${SILKIT_REVISION}"
-echo
 if [ -n $SILKIT_REVISION ] ; then
-    echo "GETTING TAGS"
+    echo "WARNING: SILKIT GIT REF specified! MAKE SURE THAT THIS VERSION IS COMPATIBLE WITH THE VERSION SPECIFIED IN THE DEBIAN/UBUNTU CHANGELOG!"
+    echo "SILKIT REVISION: ${SILKIT_REVISION}"
+    echo "TRY GETTING REF AS TAGS"
     git -c http.sslVerify=false -C ./libsilkit-${SILKIT_VERSION} fetch --depth=1 origin refs/tags/${SILKIT_REVISION}:refs/tags/${SILKIT_REVISION} --no-tags
-    echo "RESETTING TO TAGS"
+    ret_val=$?
+    if [ "$ret_val" != '0' ] ; then
+        echo "TRY GETTING REF AS COMMIT SHA"
+        git -c http.sslVerify=false -C ./libsilkit-${SILKIT_VERSION} fetch --depth=1 origin ${SILKIT_REVISION}
+        ret_val=$?
+    fi
+
+    if [ "$ret_val" != '0' ] ; then
+        echo "Could not find REF $SILKIT_REVISION in $SILKIT_SOURCE_URL\nExiting"
+        exit 64
+    fi
+    echo "RESETTING TO REF: $SILKIT_REVISION"
     git -C ./libsilkit-${SILKIT_VERSION} reset --hard ${SILKIT_REVISION}
 fi
